Minor refactoring in folder hierarchy
