+++ lib/core/view_models/home/view_model.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:mobile_code_challenge_solution/core/providers/controller/provider.dart';
-import 'package:mobile_code_challenge_solution/core/view_models/base.dart';
+import 'package:mobile_code_challenge_solution/core/view_models/base/base.dart';
 
 class HomeViewModel extends BaseModel {
   HomeViewModel({
