+++ lib/core/view_models/repository_list/view_model.dart
@@ -3,7 +3,7 @@ import 'dart:collection';
 import 'package:flutter/foundation.dart';
 import 'package:mobile_code_challenge_solution/core/models/repository/repository.dart';
 import 'package:mobile_code_challenge_solution/core/providers/get_repos/provider.dart';
-import 'package:mobile_code_challenge_solution/core/view_models/base.dart';
+import 'package:mobile_code_challenge_solution/core/view_models/base/base.dart';
 
 class RepositoryListViewModel extends BaseModel {
   RepositoryListViewModel({
