+++ lib/ui/views/home/view.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:mobile_code_challenge_solution/core/providers/controller/provider.dart';
 import 'package:mobile_code_challenge_solution/core/view_models/home/view_model.dart';
-import 'package:mobile_code_challenge_solution/ui/views/base.dart';
+import 'package:mobile_code_challenge_solution/ui/views/base/base.dart';
 import 'package:mobile_code_challenge_solution/ui/views/repository_list/view.dart';
 import 'package:mobile_code_challenge_solution/ui/widgets/bottom_nav.dart';
 import 'package:provider/provider.dart';
