+++ lib/ui/views/repository_list/view.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:loadmore/loadmore.dart';
 import 'package:mobile_code_challenge_solution/core/providers/get_repos/provider.dart';
 import 'package:mobile_code_challenge_solution/core/view_models/repository_list/view_model.dart';
-import 'package:mobile_code_challenge_solution/ui/views/base.dart';
+import 'package:mobile_code_challenge_solution/ui/views/base/base.dart';
 import 'package:mobile_code_challenge_solution/ui/widgets/repository_item.dart';
 import 'package:provider/provider.dart';
